adjusting position of weird custom font icons
ok @kmcgillivray LOL so i know this is messy. i labeled though!!! ARENT
YOU PROUD!? so obviously you’ll need to go in here and check out how I
actually did it, and maybe see if there is a more sustainable way to do
it. but at least you can see what it is i wanted it to look like.
Basically just aligning the icons with the parenthesis correctly, and
making sure they are similar size.  I couldn’t quite figure out how to
do the spacing of the elements up there, and the animations for the
parenthesis perfectly, or would be open to you changing both of those
things!